- Separated out smoke and smokePerf test
